Do refactor to make build pass
